# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing weather subscriptions and fetching weather data.",
        "title": "Weather Subscriptions API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/confirm/{token}": {
            "get": {
                "description": "Confirms a subscription using the token sent in the confirmation email.",
                "tags": [
                    "subscriptions"
                ],
                "summary": "Confirm email subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Confirmation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscribe": {
            "post": {
                "description": "Subscribe an email to receive weather updates for a specific city and frequency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Subscribe to weather updates",
                "parameters": [
                    {
                        "description": "Subscription request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subscriptions.SubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SubscribeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/unsubscribe/{token}": {
            "get": {
                "description": "Unsubscribes an email from weather updates using the token sent in emails.",
                "tags": [
                    "subscriptions"
                ],
                "summary": "Unsubscribe from weather updates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unsubscribe token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/weather": {
            "get": {
                "description": "Returns the current weather forecast for the specified city.",
                "tags": [
                    "weather"
                ],
                "summary": "Get current weather for a city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeatherResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "city name is required"
                }
            }
        },
        "handlers.SubscribeResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "confirmation email sent"
                }
            }
        },
        "handlers.WeatherResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "partly cloudy"
                },
                "humidity": {
                    "type": "number",
                    "example": 64
                },
                "temperature": {
                    "type": "number",
                    "example": 21.5
                }
            }
        },
        "subscriptions.SubscribeRequest": {
            "type": "object",
            "required": [
                "city",
                "email",
                "frequency"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                }
            }
        }
    }
}